backend: Refactored the entire directory structure to make it more readable, clean, and compliant with django and python conventions
