+++ backend/api/services/ddb_service.py
@@ -1,4 +1,3 @@
-import json
 from django.shortcuts import render
 from django.http import JsonResponse
 from django.views.decorators.csrf import csrf_exempt
@@ -6,6 +5,9 @@ import boto3
 from boto3.dynamodb.conditions import Key, Attr
 import os
 
+# This module provides utility functions to interact with AWS DynamoDB for DICOM data retrieval.
+
+# Create DynamoDB resource
 def get_dynamodb_resource():
     return boto3.resource(
         "dynamodb",
@@ -15,7 +17,7 @@ def get_dynamodb_resource():
 @csrf_exempt
 def get_study_data_by_uid(request):
     try:
-        user_id = request.GET.get("userId")
+        user_id = request.GET.get("userId") # Get UserId from the URL query parameters
         record_type = request.GET.get("recordType")  # e.g., "study" or "instance"
 
         if not user_id or not record_type:
@@ -24,7 +26,7 @@ def get_study_data_by_uid(request):
         dynamodb_resource = get_dynamodb_resource()
         dicom_data_table = dynamodb_resource.Table("dicomFileMetadataTable")
         response = dicom_data_table.query(
-            KeyConditionExpression=Key('UserID').eq(user_id),
+            KeyConditionExpression=Key('UserId').eq(user_id),
             FilterExpression=Attr('DataType').eq(record_type)
         )
 
@@ -34,16 +36,3 @@ def get_study_data_by_uid(request):
     except Exception as e:
         print(f"Failed to get DICOM data: {str(e)}")
         return JsonResponse({"error": f"Failed to get DICOM data: {str(e)}"}, status=500)
-
-# @csrf_exempt
-# def get_study_data_by_uid(request):
-#     try:
-#         dicom_data_table = dynamodb.Table("dicomFileMetadataTable")
-#         response = dicom_data_table.query(
-#         KeyConditionExpression=boto3.dynamodb.conditions.Key('UserID').eq(request.GET.get("userId"))
-#         )
-#         items = response['Items']
-#         return JsonResponse(items, safe=False)
-#     except Exception as e:
-#             print(f"Failed to get DICOM data: {str(e)}")
-#             return JsonResponse({"error": f"Failed to get Dicom Data: {str(e)}"}, status=500)
