--- backend/api/try.py
@@ -1,27 +1,0 @@
-import boto3
-from boto3.dynamodb.conditions import Key
-import os
-from .ddb_utils import get_dynamodb_resource
-
-dynamodb = get_dynamodb_resource()
-table = dynamodb.Table('dicomFileMetadataTable')
-
-response = table.query(
-            IndexName='StudyInstanceUID-SeriesInstanceUID-index',
-            KeyConditionExpression=Key('StudyInstanceUID').eq("2.25.332499716250305887681441969149281156598")
-)
-
-print(response)
-
-# {'Items': [{'PatientAge': '077Y', 'UploadTimestamp': '2025-04-13T09:18:24.390470', 'PatientName': 'MYERS^MARK', 'Modality': 'PT', 'HasPixelData': True, 'ImagePositionPatient': '[-250.00000000000, -250.00000000000, -866.50000000000]', 'SliceIndex': Decimal('11'), 'BitsAllocated': Decimal('16'), 'XRayTubeCurrent': 'Unknown', 'ConvolutionKernel': 'Unknown', 'ScanOptions': 'Unknown', 'SoftwareVersions': Decimal('16.01'), 'ManufacturerModelName': 'Discovery LS', 'Rows': Decimal('128'), 'StudyInstanceUID': '2.25.332499716250305887681441969149281156598', 'ExposureTime': 'Unknown', 'SliceThickness': Decimal('4.25'), 'StudyDescription': 'Unknown', 'PatientSex': 'M', 'PhotometricInterpretation': 'MONOCHROME2', 
-# 'SeriesNumber': 'Unknown', 'Manufacturer': 'GE MEDICAL SYSTEMS', 'PatientID': Decimal('6451050561'), 'FileKey': '04989458-2081-70ba-7740-7ec9c9f34b66/6451050561/2.25.332499716250305887681441969149281156598/2.25.266330404550956556941727606810649148918/2.25.125179945293118088999485642465345125878.dcm', 'DataType': 'instance', 'KVP': 'Unknown', 'InstanceNumber': Decimal('11'), 'Columns': Decimal('128'), 'ImageOrientationPatient': '[1, 0, 0, 0, 1, -0]', 'AccessionNumber': '19610729E532433', 'TotalSizeBytes': Decimal('420530'), 'ContrastBolusAgent': 'Unknown', 'SOPClassUID': '1.2.840.10008.5.1.4.1.1.128', 'UserID': '04989458-2081-70ba-7740-7ec9c9f34b66', 'WindowWidth': Decimal('400'), 'NumberOfFrames': Decimal('1'), 'PatientWeight': Decimal('78'), 'BodyPartExamined': 'Unknown', 'PixelSpacing': '[3.90625, 3.90625]', 'SeriesDescription': 'PET IR CTAC WB', 'BitsStored': Decimal('16'), 'StudyID': 'Unknown', 'FrameOfReferenceUID': '2.25.86249723901431855689260781743269068278', 'SOPInstanceUID': '2.25.125179945293118088999485642465345125878', 'StudyDate': Decimal('19610728'), 'WindowCenter': Decimal('40'), 'SeriesInstanceUID': '2.25.266330404550956556941727606810649148918'}, {'PatientAge': '077Y', 'UploadTimestamp': '2025-04-13T09:18:24.390470', 'PatientName': 'MYERS^MARK', 'Modality': 'PT', 'HasPixelData': True, 'ImagePositionPatient': '[-250.00000000000, -250.00000000000, -904.75000000000]', 'SliceIndex': 
-# Decimal('2'), 'BitsAllocated': Decimal('16'), 'XRayTubeCurrent': 'Unknown', 'ConvolutionKernel': 'Unknown', 'ScanOptions': 'Unknown', 'SoftwareVersions': Decimal('16.01'), 'ManufacturerModelName': 'Discovery LS', 'Rows': Decimal('128'), 'StudyInstanceUID': '2.25.332499716250305887681441969149281156598', 'ExposureTime': 'Unknown', 'SliceThickness': Decimal('4.25'), 'StudyDescription': 'Unknown', 'PatientSex': 'M', 'PhotometricInterpretation': 'MONOCHROME2', 'SeriesNumber': 'Unknown', 'Manufacturer': 'GE MEDICAL SYSTEMS', 'PatientID': Decimal('6451050561'), 'FileKey': '04989458-2081-70ba-7740-7ec9c9f34b66/6451050561/2.25.332499716250305887681441969149281156598/2.25.266330404550956556941727606810649148918/2.25.163282943459288825504287870963364249078.dcm', 'DataType': 'instance', 'KVP': 'Unknown', 'InstanceNumber': Decimal('2'), 'Columns': Decimal('128'), 'ImageOrientationPatient': '[1, 0, 0, 0, 1, -0]', 'AccessionNumber': '19610729E532433', 'TotalSizeBytes': Decimal('76460'), 'ContrastBolusAgent': 'Unknown', 'SOPClassUID': '1.2.840.10008.5.1.4.1.1.128', 'UserID': '04989458-2081-70ba-7740-7ec9c9f34b66', 'WindowWidth': Decimal('400'), 'NumberOfFrames': Decimal('1'), 'PatientWeight': Decimal('78'), 'BodyPartExamined': 'Unknown', 'PixelSpacing': '[3.90625, 3.90625]', 'SeriesDescription': 'PET IR CTAC WB', 'BitsStored': Decimal('16'), 'StudyID': 'Unknown', 'FrameOfReferenceUID': '2.25.86249723901431855689260781743269068278', 'SOPInstanceUID': '2.25.163282943459288825504287870963364249078', 'StudyDate': Decimal('19610728'), 'WindowCenter': Decimal('40'), 'SeriesInstanceUID': '2.25.266330404550956556941727606810649148918'}, {'PatientAge': '077Y', 'UploadTimestamp': '2025-04-13T09:18:24.390470', 'PatientName': 'MYERS^MARK', 'Modality': 'PT', 'HasPixelData': True, 'ImagePositionPatient': '[-250.00000000000, -250.00000000000, -870.75000000000]', 'SliceIndex': Decimal('10'), 'BitsAllocated': Decimal('16'), 'XRayTubeCurrent': 'Unknown', 'ConvolutionKernel': 'Unknown', 
-# 'ScanOptions': 'Unknown', 'SoftwareVersions': Decimal('16.01'), 'ManufacturerModelName': 'Discovery LS', 'Rows': Decimal('128'), 'StudyInstanceUID': '2.25.332499716250305887681441969149281156598', 'ExposureTime': 'Unknown', 'SliceThickness': Decimal('4.25'), 'StudyDescription': 'Unknown', 'PatientSex': 'M', 'PhotometricInterpretation': 'MONOCHROME2', 'SeriesNumber': 'Unknown', 'Manufacturer': 'GE MEDICAL SYSTEMS', 'PatientID': Decimal('6451050561'), 'FileKey': '04989458-2081-70ba-7740-7ec9c9f34b66/6451050561/2.25.332499716250305887681441969149281156598/2.25.266330404550956556941727606810649148918/2.25.152797016922363426159444735592444313078.dcm', 'DataType': 'instance', 'KVP': 'Unknown', 'InstanceNumber': Decimal('10'), 'Columns': Decimal('128'), 'ImageOrientationPatient': '[1, 0, 0, 0, 1, -0]', 'AccessionNumber': '19610729E532433', 'TotalSizeBytes': Decimal('382300'), 'ContrastBolusAgent': 'Unknown', 'SOPClassUID': '1.2.840.10008.5.1.4.1.1.128', 'UserID': '04989458-2081-70ba-7740-7ec9c9f34b66', 'WindowWidth': Decimal('400'), 'NumberOfFrames': Decimal('1'), 'PatientWeight': Decimal('78'), 'BodyPartExamined': 'Unknown', 'PixelSpacing': '[3.90625, 3.90625]', 'SeriesDescription': 'PET IR CTAC WB', 'BitsStored': Decimal('16'), 'StudyID': 'Unknown', 'FrameOfReferenceUID': '2.25.86249723901431855689260781743269068278', 'SOPInstanceUID': '2.25.152797016922363426159444735592444313078', 'StudyDate': Decimal('19610728'), 'WindowCenter': Decimal('40'), 'SeriesInstanceUID': '2.25.266330404550956556941727606810649148918'}, {'PatientAge': '077Y', 'UploadTimestamp': '2025-04-13T09:18:24.390470', 'PatientName': 'MYERS^MARK', 'Modality': 'PT', 'HasPixelData': True, 'ImagePositionPatient': '[-250.00000000000, -250.00000000000, -887.75000000000]', 'SliceIndex': Decimal('6'), 'BitsAllocated': Decimal('16'), 'XRayTubeCurrent': 'Unknown', 'ConvolutionKernel': 'Unknown', 'ScanOptions': 'Unknown', 'SoftwareVersions': Decimal('16.01'), 'ManufacturerModelName': 'Discovery LS', 'Rows': Decimal('128'), 'StudyInstanceUID': '2.25.332499716250305887681441969149281156598', 'ExposureTime': 'Unknown', 'SliceThickness': Decimal('4.25'), 'StudyDescription': 'Unknown', 'PatientSex': 'M', 'PhotometricInterpretation': 'MONOCHROME2', 'SeriesNumber': 'Unknown', 'Manufacturer': 'GE MEDICAL SYSTEMS', 'PatientID': Decimal('6451050561'), 'FileKey': '04989458-2081-70ba-7740-7ec9c9f34b66/6451050561/2.25.332499716250305887681441969149281156598/2.25.266330404550956556941727606810649148918/2.25.13998604125085923607342310500765122038.dcm', 'DataType': 'instance', 'KVP': 'Unknown', 'InstanceNumber': Decimal('6'), 'Columns': Decimal('128'), 'ImageOrientationPatient': '[1, 0, 0, 0, 1, -0]', 'AccessionNumber': '19610729E532433', 'TotalSizeBytes': Decimal('229380'), 'ContrastBolusAgent': 'Unknown', 'SOPClassUID': '1.2.840.10008.5.1.4.1.1.128', 
-# 'UserID': '04989458-2081-70ba-7740-7ec9c9f34b66', 'WindowWidth': Decimal('400'), 'NumberOfFrames': Decimal('1'), 'PatientWeight': Decimal('78'), 'BodyPartExamined': 'Unknown', 'PixelSpacing': '[3.90625, 3.90625]', 'SeriesDescription': 'PET IR CTAC WB', 'BitsStored': Decimal('16'), 'StudyID': 'Unknown', 'FrameOfReferenceUID': '2.25.86249723901431855689260781743269068278', 'SOPInstanceUID': '2.25.13998604125085923607342310500765122038', 'StudyDate': Decimal('19610728'), 'WindowCenter': Decimal('40'), 'SeriesInstanceUID': '2.25.266330404550956556941727606810649148918'}, {'PatientAge': '077Y', 'UploadTimestamp': '2025-04-13T09:18:24.390470', 'PatientName': 'MYERS^MARK', 'Modality': 'PT', 'HasPixelData': True, 'ImagePositionPatient': '[-250.00000000000, -250.00000000000, -896.25000000000]', 'SliceIndex': Decimal('4'), 'BitsAllocated': Decimal('16'), 'XRayTubeCurrent': 'Unknown', 'ConvolutionKernel': 'Unknown', 'ScanOptions': 'Unknown', 'SoftwareVersions': Decimal('16.01'), 'ManufacturerModelName': 'Discovery LS', 'Rows': Decimal('128'), 'StudyInstanceUID': '2.25.332499716250305887681441969149281156598', 'ExposureTime': 'Unknown', 'SliceThickness': Decimal('4.25'), 'StudyDescription': 'Unknown', 'PatientSex': 'M', 'PhotometricInterpretation': 'MONOCHROME2', 'SeriesNumber': 'Unknown', 'Manufacturer': 'GE MEDICAL SYSTEMS', 'PatientID': Decimal('6451050561'), 'FileKey': '04989458-2081-70ba-7740-7ec9c9f34b66/6451050561/2.25.332499716250305887681441969149281156598/2.25.266330404550956556941727606810649148918/2.25.122386122598377585662924502144646757878.dcm', 'DataType': 'instance', 'KVP': 'Unknown', 'InstanceNumber': Decimal('4'), 'Columns': Decimal('128'), 'ImageOrientationPatient': '[1, 0, 0, 0, 1, -0]', 'AccessionNumber': '19610729E532433', 'TotalSizeBytes': Decimal('152920'), 'ContrastBolusAgent': 'Unknown', 'SOPClassUID': '1.2.840.10008.5.1.4.1.1.128', 'UserID': '04989458-2081-70ba-7740-7ec9c9f34b66', 'WindowWidth': Decimal('400'), 'NumberOfFrames': Decimal('1'), 'PatientWeight': Decimal('78'), 'BodyPartExamined': 'Unknown', 'PixelSpacing': '[3.90625, 3.90625]', 'SeriesDescription': 'PET IR CTAC WB', 'BitsStored': Decimal('16'), 'StudyID': 'Unknown', 'FrameOfReferenceUID': '2.25.86249723901431855689260781743269068278', 'SOPInstanceUID': '2.25.122386122598377585662924502144646757878', 'StudyDate': Decimal('19610728'), 'WindowCenter': Decimal('40'), 'SeriesInstanceUID': '2.25.266330404550956556941727606810649148918'}, {'PatientAge': '077Y', 'UploadTimestamp': '2025-04-13T09:18:24.390470', 'PatientName': 'MYERS^MARK', 'Modality': 'PT', 'HasPixelData': True, 'ImagePositionPatient': '[-250.00000000000, -250.00000000000, -909.00000000000]', 'SliceIndex': Decimal('1'), 'BitsAllocated': Decimal('16'), 'XRayTubeCurrent': 'Unknown', 'ConvolutionKernel': 'Unknown', 'ScanOptions': 'Unknown', 'SoftwareVersions': Decimal('16.01'), 'ManufacturerModelName': 'Discovery LS', 'Rows': Decimal('128'), 'StudyInstanceUID': '2.25.332499716250305887681441969149281156598', 'ExposureTime': 'Unknown', 'SliceThickness': Decimal('4.25'), 'StudyDescription': 'Unknown', 'PatientSex': 'M', 'PhotometricInterpretation': 'MONOCHROME2', 'SeriesNumber': 'Unknown', 'Manufacturer': 'GE MEDICAL SYSTEMS', 'PatientID': Decimal('6451050561'), 'FileKey': '04989458-2081-70ba-7740-7ec9c9f34b66/6451050561/2.25.332499716250305887681441969149281156598/2.25.266330404550956556941727606810649148918/2.25.53276105529239306645897576484909465078.dcm', 'DataType': 'instance', 'KVP': 'Unknown', 'InstanceNumber': Decimal('1'), 'Columns': Decimal('128'), 'ImageOrientationPatient': '[1, 0, 0, 0, 1, -0]', 'AccessionNumber': '19610729E532433', 'TotalSizeBytes': Decimal('38230'), 'ContrastBolusAgent': 'Unknown', 'SOPClassUID': '1.2.840.10008.5.1.4.1.1.128', 'UserID': '04989458-2081-70ba-7740-7ec9c9f34b66', 'WindowWidth': Decimal('400'), 'NumberOfFrames': Decimal('1'), 'PatientWeight': Decimal('78'), 'BodyPartExamined': 'Unknown', 'PixelSpacing': '[3.90625, 3.90625]', 'SeriesDescription': 'PET IR CTAC WB', 'BitsStored': Decimal('16'), 'StudyID': 'Unknown', 'FrameOfReferenceUID': '2.25.86249723901431855689260781743269068278', 'SOPInstanceUID': '2.25.53276105529239306645897576484909465078', 'StudyDate': Decimal('19610728'), 'WindowCenter': Decimal('40'), 'SeriesInstanceUID': '2.25.266330404550956556941727606810649148918'}, {'PatientAge': '077Y', 'UploadTimestamp': '2025-04-13T09:18:24.390470', 'PatientName': 'MYERS^MARK', 'Modality': 'PT', 'HasPixelData': True, 'ImagePositionPatient': '[-250.00000000000, -250.00000000000, -883.50000000000]', 'SliceIndex': Decimal('7'), 'BitsAllocated': Decimal('16'), 'XRayTubeCurrent': 'Unknown', 'ConvolutionKernel': 'Unknown', 'ScanOptions': 'Unknown', 'SoftwareVersions': Decimal('16.01'), 'ManufacturerModelName': 'Discovery LS', 'Rows': Decimal('128'), 'StudyInstanceUID': '2.25.332499716250305887681441969149281156598', 'ExposureTime': 'Unknown', 'SliceThickness': Decimal('4.25'), 'StudyDescription': 'Unknown', 'PatientSex': 'M', 'PhotometricInterpretation': 'MONOCHROME2', 'SeriesNumber': 'Unknown', 'Manufacturer': 'GE MEDICAL SYSTEMS', 'PatientID': Decimal('6451050561'), 'FileKey': '04989458-2081-70ba-7740-7ec9c9f34b66/6451050561/2.25.332499716250305887681441969149281156598/2.25.266330404550956556941727606810649148918/2.25.324768301374585476565454497429501623798.dcm', 'DataType': 'instance', 'KVP': 'Unknown', 'InstanceNumber': Decimal('7'), 'Columns': Decimal('128'), 'ImageOrientationPatient': '[1, 0, 0, 0, 1, -0]', 'AccessionNumber': '19610729E532433', 'TotalSizeBytes': 
-# Decimal('267610'), 'ContrastBolusAgent': 'Unknown', 'SOPClassUID': '1.2.840.10008.5.1.4.1.1.128', 'UserID': '04989458-2081-70ba-7740-7ec9c9f34b66', 'WindowWidth': Decimal('400'), 'NumberOfFrames': Decimal('1'), 'PatientWeight': Decimal('78'), 'BodyPartExamined': 'Unknown', 'PixelSpacing': '[3.90625, 3.90625]', 'SeriesDescription': 'PET IR CTAC WB', 'BitsStored': Decimal('16'), 'StudyID': 'Unknown', 'FrameOfReferenceUID': '2.25.86249723901431855689260781743269068278', 'SOPInstanceUID': '2.25.324768301374585476565454497429501623798', 'StudyDate': Decimal('19610728'), 'WindowCenter': Decimal('40'), 'SeriesInstanceUID': '2.25.266330404550956556941727606810649148918'}, {'PatientAge': '077Y', 'UploadTimestamp': '2025-04-13T09:18:24.390470', 'PatientName': 'MYERS^MARK', 'Modality': 'PT', 'HasPixelData': True, 'ImagePositionPatient': '[-250.00000000000, -250.00000000000, -900.50000000000]', 'SliceIndex': Decimal('3'), 'BitsAllocated': Decimal('16'), 'XRayTubeCurrent': 'Unknown', 'ConvolutionKernel': 'Unknown', 'ScanOptions': 'Unknown', 'SoftwareVersions': Decimal('16.01'), 'ManufacturerModelName': 'Discovery LS', 'Rows': Decimal('128'), 'StudyInstanceUID': '2.25.332499716250305887681441969149281156598', 'ExposureTime': 'Unknown', 'SliceThickness': Decimal('4.25'), 'StudyDescription': 'Unknown', 'PatientSex': 'M', 'PhotometricInterpretation': 'MONOCHROME2', 'SeriesNumber': 'Unknown', 'Manufacturer': 'GE MEDICAL SYSTEMS', 'PatientID': Decimal('6451050561'), 'FileKey': '04989458-2081-70ba-7740-7ec9c9f34b66/6451050561/2.25.332499716250305887681441969149281156598/2.25.266330404550956556941727606810649148918/2.25.329688616270165898720781804699333017078.dcm', 'DataType': 'instance', 'KVP': 
-# 'Unknown', 'InstanceNumber': Decimal('3'), 'Columns': Decimal('128'), 'ImageOrientationPatient': '[1, 0, 0, 0, 1, -0]', 'AccessionNumber': '19610729E532433', 'TotalSizeBytes': Decimal('114690'), 'ContrastBolusAgent': 'Unknown', 'SOPClassUID': '1.2.840.10008.5.1.4.1.1.128', 'UserID': 
-# '04989458-2081-70ba-7740-7ec9c9f34b66', 'WindowWidth': Decimal('400'), 'NumberOfFrames': Decimal('1'), 'PatientWeight': Decimal('78'), 'BodyPartExamined': 'Unknown', 'PixelSpacing': '[3.90625, 3.90625]', 'SeriesDescription': 'PET IR CTAC WB', 'BitsStored': Decimal('16'), 'StudyID': 'Unknown', 'FrameOfReferenceUID': '2.25.86249723901431855689260781743269068278', 'SOPInstanceUID': '2.25.329688616270165898720781804699333017078', 'StudyDate': Decimal('19610728'), 'WindowCenter': Decimal('40'), 'SeriesInstanceUID': '2.25.266330404550956556941727606810649148918'}, {'PatientAge': '077Y', 'UploadTimestamp': '2025-04-13T09:18:24.390470', 'PatientName': 'MYERS^MARK', 'Modality': 'PT', 'HasPixelData': True, 'ImagePositionPatient': '[-250.00000000000, -250.00000000000, -892.00000000000]', 'SliceIndex': Decimal('5'), 'BitsAllocated': Decimal('16'), 'XRayTubeCurrent': 'Unknown', 'ConvolutionKernel': 'Unknown', 'ScanOptions': 'Unknown', 'SoftwareVersions': Decimal('16.01'), 'ManufacturerModelName': 'Discovery LS', 'Rows': Decimal('128'), 'StudyInstanceUID': '2.25.332499716250305887681441969149281156598', 'ExposureTime': 'Unknown', 'SliceThickness': Decimal('4.25'), 'StudyDescription': 'Unknown', 'PatientSex': 'M', 'PhotometricInterpretation': 'MONOCHROME2', 'SeriesNumber': 'Unknown', 'Manufacturer': 'GE MEDICAL SYSTEMS', 'PatientID': Decimal('6451050561'), 'FileKey': '04989458-2081-70ba-7740-7ec9c9f34b66/6451050561/2.25.332499716250305887681441969149281156598/2.25.266330404550956556941727606810649148918/2.25.101258316809433709051119099490965503478.dcm', 'DataType': 'instance', 'KVP': 'Unknown', 'InstanceNumber': Decimal('5'), 'Columns': Decimal('128'), 'ImageOrientationPatient': '[1, 0, 0, 
-# 0, 1, -0]', 'AccessionNumber': '19610729E532433', 'TotalSizeBytes': Decimal('191150'), 'ContrastBolusAgent': 'Unknown', 'SOPClassUID': '1.2.840.10008.5.1.4.1.1.128', 'UserID': '04989458-2081-70ba-7740-7ec9c9f34b66', 'WindowWidth': Decimal('400'), 'NumberOfFrames': Decimal('1'), 'PatientWeight': Decimal('78'), 'BodyPartExamined': 'Unknown', 'PixelSpacing': '[3.90625, 3.90625]', 'SeriesDescription': 'PET IR CTAC WB', 'BitsStored': Decimal('16'), 'StudyID': 'Unknown', 'FrameOfReferenceUID': '2.25.86249723901431855689260781743269068278', 'SOPInstanceUID': '2.25.101258316809433709051119099490965503478', 'StudyDate': Decimal('19610728'), 'WindowCenter': Decimal('40'), 'SeriesInstanceUID': '2.25.266330404550956556941727606810649148918'}, {'PatientAge': '077Y', 'UploadTimestamp': '2025-04-13T09:18:24.390470', 'PatientName': 'MYERS^MARK', 'Modality': 'PT', 'HasPixelData': True, 'ImagePositionPatient': '[-250.00000000000, -250.00000000000, -879.25000000000]', 'SliceIndex': Decimal('8'), 'BitsAllocated': Decimal('16'), 'XRayTubeCurrent': 'Unknown', 'ConvolutionKernel': 'Unknown', 'ScanOptions': 'Unknown', 'SoftwareVersions': Decimal('16.01'), 'ManufacturerModelName': 'Discovery LS', 'Rows': Decimal('128'), 'StudyInstanceUID': '2.25.332499716250305887681441969149281156598', 'ExposureTime': 'Unknown', 'SliceThickness': Decimal('4.25'), 'StudyDescription': 'Unknown', 'PatientSex': 'M', 'PhotometricInterpretation': 'MONOCHROME2', 'SeriesNumber': 'Unknown', 'Manufacturer': 'GE MEDICAL SYSTEMS', 'PatientID': Decimal('6451050561'), 'FileKey': '04989458-2081-70ba-7740-7ec9c9f34b66/6451050561/2.25.332499716250305887681441969149281156598/2.25.266330404550956556941727606810649148918/2.25.96600500265754734927785034069667593718.dcm', 'DataType': 'instance', 'KVP': 'Unknown', 'InstanceNumber': Decimal('8'), 'Columns': Decimal('128'), 'ImageOrientationPatient': '[1, 0, 0, 0, 1, -0]', 'AccessionNumber': '19610729E532433', 'TotalSizeBytes': Decimal('305840'), 'ContrastBolusAgent': 
-# 'Unknown', 'SOPClassUID': '1.2.840.10008.5.1.4.1.1.128', 'UserID': '04989458-2081-70ba-7740-7ec9c9f34b66', 'WindowWidth': Decimal('400'), 'NumberOfFrames': Decimal('1'), 'PatientWeight': Decimal('78'), 'BodyPartExamined': 'Unknown', 'PixelSpacing': '[3.90625, 3.90625]', 'SeriesDescription': 'PET IR CTAC WB', 'BitsStored': Decimal('16'), 'StudyID': 'Unknown', 'FrameOfReferenceUID': '2.25.86249723901431855689260781743269068278', 'SOPInstanceUID': '2.25.96600500265754734927785034069667593718', 'StudyDate': Decimal('19610728'), 'WindowCenter': Decimal('40'), 'SeriesInstanceUID': '2.25.266330404550956556941727606810649148918'}, {'PatientAge': '077Y', 'UploadTimestamp': '2025-04-13T09:18:24.390470', 'PatientName': 'MYERS^MARK', 'Modality': 'PT', 'StudyUIDHash': 'd9d671f91623bbb09a479bb76baab9dc86370a93', 'XRayTubeCurrent': 'Unknown', 'ConvolutionKernel': 'Unknown', 'ScanOptions': 'Unknown', 'SoftwareVersions': Decimal('16.01'), 'ManufacturerModelName': 'Discovery LS', 'StudyInstanceUID': '2.25.332499716250305887681441969149281156598', 'ExposureTime': 'Unknown', 'TotalStudySizeBytes': Decimal('458760'), 'StudyDescription': 'Unknown', 'PatientSex': 'M', 'SeriesInstanceUIDList': ['2.25.266330404550956556941727606810649148918'], 'SeriesNumber': 'Unknown', 'Manufacturer': 'GE MEDICAL SYSTEMS', 'StudyName': 'Unknown', 'PatientID': Decimal('6451050561'), 'FileKey': '04989458-2081-70ba-7740-7ec9c9f34b66/6451050561/2.25.332499716250305887681441969149281156598/', 'DataType': 'study', 'KVP': 'Unknown', 'AccessionNumber': '19610729E532433', 'ContrastBolusAgent': 'Unknown', 'UserID': '04989458-2081-70ba-7740-7ec9c9f34b66', 'PatientWeight': Decimal('78'), 'BodyPartExamined': 'Unknown', 'SeriesDescription': 'PET IR CTAC WB', 'StudyID': 'Unknown', 'SOPInstanceUIDList': ['2.25.53276105529239306645897576484909465078', '2.25.163282943459288825504287870963364249078', '2.25.329688616270165898720781804699333017078', '2.25.122386122598377585662924502144646757878', '2.25.101258316809433709051119099490965503478', '2.25.13998604125085923607342310500765122038', '2.25.324768301374585476565454497429501623798', '2.25.96600500265754734927785034069667593718', '2.25.128634092731488853029790047033800375798', '2.25.152797016922363426159444735592444313078', '2.25.125179945293118088999485642465345125878', '2.25.333011947712564485475355381045069987318'], 'StudyDate': Decimal('19610728'), 'NumberOfInstances': Decimal('12'), 'SeriesInstanceUID': '2.25.266330404550956556941727606810649148918'}, {'PatientAge': '077Y', 'UploadTimestamp': '2025-04-13T09:18:24.390470', 'PatientName': 'MYERS^MARK', 'Modality': 'PT', 'HasPixelData': True, 'ImagePositionPatient': '[-250.00000000000, -250.00000000000, -862.25000000000]', 'SliceIndex': Decimal('12'), 'BitsAllocated': Decimal('16'), 'XRayTubeCurrent': 'Unknown', 'ConvolutionKernel': 'Unknown', 'ScanOptions': 'Unknown', 'SoftwareVersions': Decimal('16.01'), 'ManufacturerModelName': 'Discovery LS', 'Rows': Decimal('128'), 'StudyInstanceUID': '2.25.332499716250305887681441969149281156598', 'ExposureTime': 'Unknown', 'SliceThickness': Decimal('4.25'), 'StudyDescription': 'Unknown', 'PatientSex': 'M', 'PhotometricInterpretation': 'MONOCHROME2', 'SeriesNumber': 'Unknown', 'Manufacturer': 'GE MEDICAL SYSTEMS', 'PatientID': Decimal('6451050561'), 'FileKey': '04989458-2081-70ba-7740-7ec9c9f34b66/6451050561/2.25.332499716250305887681441969149281156598/2.25.266330404550956556941727606810649148918/2.25.333011947712564485475355381045069987318.dcm', 'DataType': 'instance', 'KVP': 'Unknown', 'InstanceNumber': Decimal('12'), 'Columns': Decimal('128'), 'ImageOrientationPatient': '[1, 0, 0, 0, 1, -0]', 'AccessionNumber': '19610729E532433', 'TotalSizeBytes': Decimal('458760'), 'ContrastBolusAgent': 'Unknown', 'SOPClassUID': '1.2.840.10008.5.1.4.1.1.128', 'UserID': '04989458-2081-70ba-7740-7ec9c9f34b66', 'WindowWidth': Decimal('400'), 'NumberOfFrames': Decimal('1'), 'PatientWeight': Decimal('78'), 'BodyPartExamined': 'Unknown', 
-# 'PixelSpacing': '[3.90625, 3.90625]', 'SeriesDescription': 'PET IR CTAC WB', 'BitsStored': Decimal('16'), 'StudyID': 'Unknown', 'FrameOfReferenceUID': '2.25.86249723901431855689260781743269068278', 'SOPInstanceUID': '2.25.333011947712564485475355381045069987318', 'StudyDate': Decimal('19610728'), 'WindowCenter': Decimal('40'), 'SeriesInstanceUID': '2.25.266330404550956556941727606810649148918'}, {'PatientAge': '077Y', 'UploadTimestamp': '2025-04-13T09:18:24.390470', 'PatientName': 'MYERS^MARK', 'Modality': 'PT', 'HasPixelData': True, 'ImagePositionPatient': '[-250.00000000000, -250.00000000000, -875.00000000000]', 'SliceIndex': Decimal('9'), 'BitsAllocated': Decimal('16'), 'XRayTubeCurrent': 'Unknown', 'ConvolutionKernel': 'Unknown', 'ScanOptions': 'Unknown', 'SoftwareVersions': Decimal('16.01'), 'ManufacturerModelName': 'Discovery LS', 'Rows': Decimal('128'), 'StudyInstanceUID': '2.25.332499716250305887681441969149281156598', 'ExposureTime': 'Unknown', 'SliceThickness': Decimal('4.25'), 'StudyDescription': 'Unknown', 'PatientSex': 'M', 'PhotometricInterpretation': 'MONOCHROME2', 'SeriesNumber': 'Unknown', 'Manufacturer': 'GE MEDICAL SYSTEMS', 'PatientID': Decimal('6451050561'), 'FileKey': '04989458-2081-70ba-7740-7ec9c9f34b66/6451050561/2.25.332499716250305887681441969149281156598/2.25.266330404550956556941727606810649148918/2.25.128634092731488853029790047033800375798.dcm', 'DataType': 'instance', 'KVP': 'Unknown', 'InstanceNumber': Decimal('9'), 'Columns': Decimal('128'), 'ImageOrientationPatient': '[1, 0, 0, 0, 1, -0]', 'AccessionNumber': '19610729E532433', 'TotalSizeBytes': Decimal('344070'), 'ContrastBolusAgent': 'Unknown', 'SOPClassUID': '1.2.840.10008.5.1.4.1.1.128', 
-# 'UserID': '04989458-2081-70ba-7740-7ec9c9f34b66', 'WindowWidth': Decimal('400'), 'NumberOfFrames': Decimal('1'), 'PatientWeight': Decimal('78'), 'BodyPartExamined': 'Unknown', 'PixelSpacing': '[3.90625, 3.90625]', 'SeriesDescription': 'PET IR CTAC WB', 'BitsStored': Decimal('16'), 'StudyID': 'Unknown', 'FrameOfReferenceUID': '2.25.86249723901431855689260781743269068278', 'SOPInstanceUID': '2.25.128634092731488853029790047033800375798', 'StudyDate': Decimal('19610728'), 'WindowCenter': Decimal('40'), 'SeriesInstanceUID': '2.25.266330404550956556941727606810649148918'}], 'Count': 13, 'ScannedCount': 13, 'ResponseMetadata': {'RequestId': 'O7LNP6KLJ0VH8SS7MV41EJ79UFVV4KQNSO5AEMVJF66Q9ASUAAJG', 'HTTPStatusCode': 200, 'HTTPHeaders': {'server': 'Server', 'date': 'Sun, 13 Apr 2025 10:23:11 GMT', 'content-type': 'application/x-amz-json-1.0', 'content-length': '25281', 'connection': 'keep-alive', 'x-amzn-requestid': 'O7LNP6KLJ0VH8SS7MV41EJ79UFVV4KQNSO5AEMVJF66Q9ASUAAJG', 'x-amz-crc32': '2073750300'}, 'RetryAttempts': 0}}
