+++ backend/api/urls.py
@@ -1,23 +1,18 @@
 from django.urls import path
 from django.http import JsonResponse
-from .views import upload_dicom
-from .dicom_parser import parse_dicom_file
-from .ddb_utils import get_study_data_by_uid
-from .delete_utils import delete_data_by_file_key
-from .ohif_utils import generate_pre_signed_url_for_ohif
-from .ohif_utils import print_something
-from .ohif_utils import send_json_response_to_ohif
-from .stats import get_stats
+from api.views import upload_dicom
+from api.services import get_study_data_by_uid
+from api.views import delete_data_by_file_key
+from api.views import print_something
+from api.services import get_stats
 
 def api_only_root(request):
     return JsonResponse({"message": "Backend API is running."})
 
 urlpatterns = [
     path("upload-dicom", upload_dicom, name="upload-dicom"),
-    path("parse-dicom", parse_dicom_file, name="parse_dicom"),
     path("get-study-data-by-uid", get_study_data_by_uid, name="get-study-data-by-uid"),
     path("delete-data-by-file-key", delete_data_by_file_key, name="delete-data-by-file-key"),
-    path("get-ohif-response", send_json_response_to_ohif, name="get-ohif-url"),
     path("print-something", print_something, name="print-something"),
     path("stats", get_stats, name="stats"),
     path("", api_only_root)
