+++ backend/api/views/__init__.py
@@ -1,0 +1,3 @@
+from .delete import delete_data_by_file_key
+from .upload import upload_dicom
+from .print_something import print_something
